remove trailing tabs from correspondence table
hfst-fst2strings permits weights to be assigned to each entry; these
weights can be used to give preferences in the case of multiple lookups.
For example:

```
x:y	1
x:z	0
```

Here `x` transforms to `y` with weight 1 and to `z` with weight 0.
Usually the lower the weight, the more "preferred" the transformation.
This can become important with orthography normalisation: if each
correction has weight 1, then the total weight will be the number of
corrections made (a sort of edit distance). A spell-checker can use this
to select the "nearest" possible correction.

In the `correspondence` file there were tabs at the end of some lines;
this is the separator used by `hfst-fst2strings` between the
stringpair and the weight. Since there was no weight afterwards, it
complained about syntax errors in the file.

Remove the trailing tabs, since there evidently weren't any weights
intended here.